add powerbutton, refactor class function delegation

+++ dstctl.py
@@ -27,7 +27,6 @@ def iter_except(function, exception):
     except exception:
         return
 
-
 def kill_existing_server_procs():
     """Executes a Windows native 'taskkill' command targeting any dedicated server nullrenderer
     instances. Returns a string containing the commands encoded stdout."""
@@ -116,6 +115,7 @@ class Shard:
         self.is_master = self.configuration.get_typed('SHARD', 'is_master')
         self.cluster_name = cluster_name
         self.name = os.path.basename(path)
+        self.process = None
         self.command = self._get_command()
         self.output_queue = Queue(
             maxsize=1024
@@ -129,6 +129,38 @@ class Shard:
             return 'master'
         return 'slave'
     
+    def _get_command(self):
+        command = [
+            TEMP_CONST_NULLRENDERER_PATH,
+            "-console_enabled",
+            "-cluster",
+            self.cluster_name,
+            "-shard",
+            self.name,
+        ]        
+        return command
+
+    def _update_output(self, q):
+        """Adds stdout of associated process to shard's output queue."""
+        if self.is_running:
+            try:
+                with self.process.stdout as pipe:
+                    for line in iter(pipe.readline, b""):
+                        q.put(line)
+            finally:
+                q.put(None)
+
+    def get_output(self):  # Schedule me
+        """Returns output queue of the shard."""
+        if self.is_running:
+            for line in iter_except(self.q.get_nowait, Empty):  # display all
+                if line is None:
+                    return None
+                else:
+                    return line
+        else:
+            return None
+
     def start(self):
         cwd = os.path.realpath(
             TEMP_CONST_STEAMCMD_DST_BIN_PATH
@@ -144,39 +176,9 @@ class Shard:
         self.thread_reader.daemon = True
         self.thread_reader.start()
            
-    def _get_command(self):
-        command = [
-            TEMP_CONST_NULLRENDERER_PATH,
-            "-console_enabled",
-            "-cluster",
-            self.cluster_name,
-            "-shard",
-            self.name,
-        ]        
-        return command
-
-    def _update_output(self, q):
-        """Adds stdout of associated process to shard's output queue."""
-        if self.is_started():
-            try:
-                with self.process.stdout as pipe:
-                    for line in iter(pipe.readline, b""):
-                        q.put(line)
-            finally:
-                q.put(None)
-
-    def get_output(self):  # Schedule me
-        """Returns output queue of the shard."""
-        if self.is_started():
-            for line in iter_except(self.q.get_nowait, Empty):  # display all
-                if line is None:
-                    return None
-                else:
-                    return line
-
     def write_input(self, line):
         """Writes the input queue of the shard to the stdin of its associated process."""
-        if self.is_started():
+        if self.is_running:
             try:
                 with self.process.stdin as pipe:
                     pipe.write(line.encode())
@@ -184,41 +186,26 @@ class Shard:
             except:
                 print(line + " NOT recieved by " + self.name)
 
-    def status(self):
-        """Returns 'UP' if associated process is found, and 'DOWN' if not."""
-        if self.is_started():
-            poll = self.process.poll()
-            if poll == None:
-                return "UP"
-            else:
-                return "DOWN"
-
-    def is_started(self):
-        return hasattr(self, "process")
-
-    def get_pid(self):
-        if self.is_started():
-            try:
-                return self.process.pid
-            except:
-                print('Could not retrieve pid for shard ' + self.name)
-        else:
-            print('Shard ' + self.name + ' is not started. A pid could not be retrieved.')
+    def get_process_id(self):
+        return self.process.pid
 
     def kill_process(self):
-        try:
-            pid = self.get_pid()
+        if self.is_running:
             self.process.kill()
-        except:
-            print('Could not kill process associated with shard: ' + self.name + '. (pid: ' + str(pid) + ")")
-        else:
-            print('Successfully killed process associated with shard: ' + self.name + '. (pid: ' + str(pid) + ")")
+    @property
+    def is_running(self):
+        if self.process != None:
+            poll = self.process.poll()
+            if poll == None:
+                return True
+            else:
+                return False
 
 
 class DedicatedServer:
-    """Represents a server cluster instance. Has methods and properties related to its folder on disk,
+    """Represents a server cluster configuration and fileset. Has methods and properties related to its folder on disk,
     its configuration (cluster.ini), and is parent to one or multiple Shards. Is the high-level container
-    object controlled by the ServerControl class."""
+    object loaded by the ServerControl class."""
 
     class NonexistentDirectory(Exception):
         pass
@@ -245,19 +232,7 @@ class DedicatedServer:
             shards.append(shard)
         return shards
 
-    @property
-    def master(self):
-        for shard in self.shards:
-            if shard.is_master:
-                return shard
-    
-    @property
-    def slave(self):
-        for shard in self.shards:
-            if not shard.is_master:
-                return shard
-        
-    def _detect_shard_directories(self): #TODO: maybe move
+    def _detect_shard_directories(self):
         shard_directories = []
         for directory in os.listdir((self.path)):
             path = os.path.join(self.path, directory)
@@ -285,39 +260,56 @@ class DedicatedServer:
                         if "server.ini" in os.listdir(full_member_path):
                             return True
         return False
-    # def add_shard(self, path):
+    @property
+    def master(self):
+        for shard in self.shards:
+            if shard.is_master:
+                return shard
+    @property
+    def slave(self):
+        for shard in self.shards:
+            if not shard.is_master:
+                return shard
+    @property
+    def processes(self):
+        processes = []
+        for shard in self.shards:
+            if shard.is_running:
+                processes.append(shard.process)
+        return processes
 
     def start(self):
         for shard in self.shards:
-            shard.start()
-
-    def shutdown(self):
-        for shard in self.shards:
-            shard.write_input('c_shutdown()')
-
-    def kill_processes(self):
-        for shard in self.shards:
-            shard.kill_process()
-            print('SHARD ' + shard.name + ' KILLED')
-
+            if not shard.is_running:
+                shard.start()
+        
     def run_command(self, command):
         for shard in self.shards:
             shard.write_input(command)
+
+    def kill(self):
+        for shard in self.shards:
+            if shard.is_running:
+                shard.kill_process()
 
 
 class ServerControl:
     """Main application. Includes the window and its components. Displays output from shards and is host to various server controls. Parent to a DedicatedServer object."""
 
-    def __init__(self, root, server=None):
+    def __init__(self, root):
         self.window = root
         self.initialize_ui()
-        self.active_server = server
+        self.active_server = None
+
+        if self.active_server:
+            print('')
 
         self.update_ui()
 
     def initialize_ui(self):
         """Setup widgets and styling of main window."""
 
+        # Window
         self.window.geometry("960x650")
         self.window.resizable(0, 0)
         self.window.protocol("WM_DELETE_WINDOW", self.quit)
@@ -326,27 +318,15 @@ class ServerControl:
         self.window.iconbitmap(icon_path)
         self.window.configure(bg="#424242")
 
-        # Style
-        style = ttk.Style()
+        # Styling
         text_font = ("TkDefaultFont", "14")
         self.window.option_add("*TButton*Label.font", text_font)
         self.window.option_add("*TCombobox*Listbox.background", "#424242")
-        self.window.option_add("*TCombobox*Listbox.selectBackground", "#525252")
-        
-        # Menubar
-        menu_bar = widgets.WidgetMenuBar(
-            master=self.window,
-            structure={
-                "File" : {
-                    "Browse" : self.on_browse
-                }
-            }
-        )
-        self.window.config(menu=menu_bar)
-        
+        self.window.option_add("*TCombobox*Listbox.selectBackground", "#525252")       
+
         # Top Bar
         self.frmTopBar = ttk.Frame(root, height=20, width=761)
-        self.frmTopBar.place(x=20, y=10)
+        self.frmTopBar.place(x=20, y=25)
 
         self.entServer = ttk.Entry(self.frmTopBar, width=50)
         self.entServer.grid(row=0, column=0)
@@ -356,28 +336,22 @@ class ServerControl:
         
         self.btnConfigureServer = ttk.Button(master=self.frmTopBar, command=self.on_configure, text="Configure")
         self.btnConfigureServer.grid(row=0, column=2)
+
+        # Power Button
+        self.btn_power = widgets.WidgetPowerButton(root, command=self.on_power)
+        self.btn_power.place(x=840,y=8)
+
         # Console Views
         self.console_view_master = widgets.WidgetConsoleView(self.window, width=450, height_in_rows=21)
-        self.console_view_master.place(x=20,y=50)
+        self.console_view_master.place(x=20,y=75)
 
         self.console_view_slave = widgets.WidgetConsoleView(self.window, width=450, height_in_rows=21)
-        self.console_view_slave.place(x=490,y=50)
-        # Status Labels
-        style.configure('status.Label', background="#353535")
-        self.lblMasterStatus = ttk.Label(root, text="STATUS: ", style='status.Label')
-        self.lblMasterStatus.update()
-        self.lblMasterStatus.place(x=20, y=(self.console_view_master.winfo_y() + self.console_view_master.height_in_pixels - 15))
-
-        self.lblSlaveStatus = ttk.Label(root, text="STATUS: ", style='status.Label')
-        self.lblSlaveStatus.update()
-        self.lblSlaveStatus.place(x=490, y=(self.console_view_master.winfo_y() + self.console_view_master.height_in_pixels - 15))
+        self.console_view_slave.place(x=490,y=75)
 
         # Quick Commands
         self.command_panel = widgets.WidgetCommandPanel(
             parent=self.window,
             buttons=[
-                ("Start", self.on_start),
-                ("Shutdown", self.on_shutdown),
                 ("Regenerate", self.on_regenerate),
                 ("Reset", self.on_reset),
                 ("Save", self.on_save),
@@ -399,36 +373,43 @@ class ServerControl:
                 ("PVP Enabled:", self.get_server_info, ['GAMEPLAY', 'pvp']),
                 ("Gamemode:", self.get_server_info, ['GAMEPLAY', 'game_mode'])
             ],
-            header='Info'
+            header='Server Info'
         )
-        self.info_panel_server.place(x=85,y=520)
+        self.info_panel_server.place(x=50,y=550,width=300)
 
     def update_ui(self):
         """Retrieves info from the active server and its shards for display in the UI.
         Includes: Shard (nullrenderer) output into ConsoleViews.
         Self-reschedules for 40ms effective intervals """
         try:
-            if self.active_server: # Update ConsoleViews & status labels
+            if self.active_server: # Update ConsoleViews & button states    
+                self.btn_power.set_state('normal')
                 self.update_console_view(self.console_view_master, self.active_server.master)
-                self.lblMasterStatus.configure(text="Status: " + str(self.active_server.master.status()))
                 self.update_console_view(self.console_view_slave, self.active_server.slave)
-                self.lblSlaveStatus.configure(text="Status: " + str(self.active_server.slave.status()))
+            else:
+                self.btn_power.set_state('disabled')
         finally:
             self.window.after(40, self.update_ui)
 
     def update_console_view(self, view, shard):
-        if hasattr(self, 'active_server'):
-            if shard.is_started():
-                line = shard.get_output()
-                if line:
-                    view.write_line(line)
-    
-    def on_start(self):
-        for shard in self.active_server.shards:
-            shard.start()
+        if self.active_server:
+            for shard in self.active_server.shards:
+                if shard.is_running:
+                    view.set_status("Runnning (pid: " + str(shard.get_process_id()))
+                    line = shard.get_output()
+                    if line:
+                        view.write_line(line)
+                else:
+                    view.set_status(" -")
 
-    def on_shutdown(self):
-        self._send_command('c_shutdown()')
+    def on_power(self):
+        if self.active_server:
+            if len(self.active_server.processes) > 0:
+                self.active_server.run_command('c_shutdown()')
+                self.window.after(1000)
+                self.active_server.kill()
+            else:
+                self.active_server.start()
 
     def on_regenerate(self):
         self._send_command("c_regenerateworld()")
@@ -471,17 +452,16 @@ class ServerControl:
                 self.set_active_server(server)
 
     def unload_server(self):
-        if hasattr(self, 'active_server'):
-            if self.active_server is not None:
-                # Shutdown server
-                self.active_server.shutdown()
-                # Wait 2 seconds
-                self.window.after(2000)
-                # Kill procs
-                self.active_server.kill_processes()
-                # Clear console views
-                self.console_view_master.clear()
-                self.console_view_slave.clear()
+        if self.active_server != None:
+            # Shutdown server
+            self.active_server.run_command('c_shutdown()')
+            # Wait 2 seconds
+            self.window.after(2000)
+            # Kill procs
+            self.active_server.kill()
+            # Clear console views
+            self.console_view_master.clear()
+            self.console_view_slave.clear()
 
     def set_active_server(self, server): 
         """Updates active_server attribute from server and updates various UI elements accordingly."""
